게시판 개선 및 예외처리(중)

+++ semi/src/main/java/com/spring/semi/controller/InfoBoardController.java
@@ -15,6 +15,7 @@ import com.spring.semi.dao.BoardDao;
 import com.spring.semi.dao.MemberDao;
 import com.spring.semi.dto.BoardDto;
 import com.spring.semi.dto.MemberDto;
+import com.spring.semi.error.TargetNotfoundException;
 
 import jakarta.servlet.http.HttpSession;
 
@@ -56,6 +57,7 @@ public class InfoBoardController {
 	@RequestMapping("/detail")
 	public String detail(Model model, @RequestParam int boardNo) {
 		BoardDto boardDto = boardDao.selectOne(boardNo);
+		if(boardDto == null) throw new TargetNotfoundException("존재하지 않는 글");
 		model.addAttribute("boardDto", boardDto);
 		
 		if(boardDto.getBoardWriter() != null) {
@@ -65,22 +67,19 @@ public class InfoBoardController {
 		return "/WEB-INF/views/infoBoard/detail.jsp";				
 	}
 	//삭제
-	@GetMapping("/delete")
-	public String showDeleteForm(@RequestParam int boardNo, Model model) {
-	    BoardDto boardDto = boardDao.selectOne(boardNo);
-	    model.addAttribute("boardDto", boardDto);
-	    return "/WEB-INF/views/infoBoard/delete.jsp";
-	}
-	@PostMapping("/delete")
-	public String doDelete(@RequestParam int boardNo) {
-	    boardDao.delete(boardNo);
-	    return "redirect:list";
+	@RequestMapping("/delete")
+	public String delete(@RequestParam int boardNo) {
+		BoardDto boardDto = boardDao.selectOne(boardNo);
+		if(boardDto == null) throw new TargetNotfoundException("존재하지 않는 글");
+		boardDao.delete(boardNo);
+		return "redirect:list";
 	}
 
 	//수정
 	@GetMapping("/edit")
 	public String edit(Model model, @RequestParam int boardNo) {
-		BoardDto boardDto = boardDao.selectOne(boardNo);//후에 예외처리를 위한
+		BoardDto boardDto = boardDao.selectOne(boardNo);
+		if(boardDto == null) throw new TargetNotfoundException("존재하지 않는 글");
 		model.addAttribute("boardDto", boardDto);
 		return "/WEB-INF/views/infoBoard/edit.jsp";
 	}
